requirements.txt
blis==0.7.8
catalogue==2.0.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cymem==2.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.7
nltk==3.7
numpy==1.23.1
packaging==21.3
pathy==0.6.2
preshed==3.0.6
pydantic==1.9.1
pyparsing==3.0.9
python-slugify==6.1.2
regex==2022.7.9
requests==2.28.1
smart-open==5.2.1
spacy==3.4.0
spacy-legacy==3.0.9
spacy-loggers==1.0.3
spacytextblob==4.0.0
srsly==2.4.4
text-unidecode==1.3
textblob==0.15.3
thinc==8.1.0
tqdm==4.64.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.10
wasabi==0.9.1
yelpapi==2.4.0
